Update cli command for 2.9 compat
