+++ ckanext/dia/plugin/pylons_plugin.py
@@ -1,7 +1,9 @@
-
 # -*- coding: utf-8 -*-
 
 import ckan.plugins as p
+
+class DIACommandsMixin(p.SingletonPlugin):
+    pass
 
 class DIANoHomepageMixin(p.SingletonPlugin):
     p.implements(p.IRoutes, inherit=True)
