+++ setup.py
@@ -30,6 +30,7 @@ setup(
     diaharvester=ckanext.dia.harvester.csw:DIASpatialHarvester
     diadcatjsonharvester=ckanext.dia.harvester.dcat:DIADCATJSONHarvester
     dianohomepage=ckanext.dia.plugin:DIANoHomepagePlugin
+    diacommands=ckanext.dia.plugin:DIACommandsPlugin
 
     [paste.paster_command]
     admin=ckanext.dia.commands:AdminCommand
